requirements.txt
Flask==3.1.0
Flask-Cors==5.0.0
tensorflow==2.16.1
pillow==11.0.0